Move disable analytics to settings
