+++ app/res/xml/commcare_preferences.xml
@@ -30,5 +30,7 @@
     <Preference
         android:key="clear-saved-session"/>
     <Preference
+        android:key="disable-analytics"/>
+    <Preference
         android:key="developer-settings"/>
 </PreferenceScreen>
